refactor: slim bootloader
